refactor(suggestions): split persistence into schema, codec and recovery
Pure relocation ahead of the durability fixes: no behaviour changes, no test
bodies touched. The old layout mixed three concerns in two files — `snapshot.rs`
and `journal.rs` each held format constants, byte translation *and* the file
reading that decides what a damaged file means — which left no obvious home for
a version-range gate or a repair policy.

Now the split is by concern rather than by record type:

- `schema.rs` — the on-disk constants both records agree on.
- `codec/` — pure bytes to data, no filesystem. One file per record type, so a
  new record lands here without touching the reader or the writer.
- `recovery.rs` — reading a store that may be damaged, in one place.
- `mod.rs` — paths and the two write paths.

Also splits `tests/persistence.rs` along the same seam: the happy path in
`round_trip`, damaged and foreign files in `recovery`.
