+++ crates/funput-suggestions/src/tests/persistence/recovery.rs
@@ -3,24 +3,8 @@ use std::io::Write;
 
 use tempfile::tempdir;
 
-use super::*;
-
-#[test]
-fn snapshot_and_journal_round_trip() {
-    let directory = tempdir().unwrap();
-    {
-        let mut engine =
-            SuggestionEngine::open(directory.path(), SuggestionConfig::default()).unwrap();
-        learned(&mut engine, "chào", 2);
-        engine.compact().unwrap();
-        learned(&mut engine, "chào", 1);
-        learned(&mut engine, "chúc", 2);
-        engine.flush().unwrap();
-    }
-    let reopened = SuggestionEngine::open(directory.path(), SuggestionConfig::default()).unwrap();
-    assert_eq!(texts(&reopened, "ch"), ["chào", "chúc"]);
-    assert_eq!(reopened.stats().words, 2);
-}
+use crate::tests::{learned, texts};
+use crate::{SuggestionConfig, SuggestionEngine};
 
 #[test]
 fn truncated_journal_keeps_complete_frames() {
@@ -78,20 +62,4 @@ fn newer_schema_is_rejected_without_overwriting_it() {
         .expect("newer schema must be unavailable");
     assert_eq!(error.kind(), std::io::ErrorKind::Unsupported);
     assert_eq!(fs::read(snapshot).unwrap(), bytes);
-}
-
-#[test]
-fn abandoned_temporary_snapshot_keeps_last_good_state() {
-    let directory = tempdir().unwrap();
-    let mut engine = SuggestionEngine::open(directory.path(), SuggestionConfig::default()).unwrap();
-    learned(&mut engine, "antoàn", 2);
-    engine.compact().unwrap();
-    fs::write(
-        directory.path().join("personal-lexicon.snapshot.tmp"),
-        b"partial",
-    )
-    .unwrap();
-    drop(engine);
-    let reopened = SuggestionEngine::open(directory.path(), SuggestionConfig::default()).unwrap();
-    assert_eq!(texts(&reopened, "an"), ["antoàn"]);
 }
